feat: response steps of main result
